+++ .vscode/settings.json
@@ -1,0 +1,10 @@
+{
+    "[rust]": {
+        "editor.defaultFormatter": "rust-lang.rust-analyzer",
+        "editor.formatOnSave": true
+    },
+    "rust-analyzer.cargo.features": [
+        "g002",
+        "virq",
+    ]
+}
