+++ interrupts.x
@@ -1,54 +1,59 @@
-/* This file should be included by hifive1 */
-/* Provide the weak symbols for each interrupt handler */
-PROVIDE(WATCHDOG = DefaultMachineExternal);
-PROVIDE(RTC = DefaultMachineExternal);
-PROVIDE(UART0 = DefaultMachineExternal);
-PROVIDE(UART1 = DefaultMachineExternal);
-PROVIDE(QSPI0 = DefaultMachineExternal);
-PROVIDE(QSPI1 = DefaultMachineExternal);
-PROVIDE(QSPI2 = DefaultMachineExternal);
-PROVIDE(GPIO0 = DefaultMachineExternal);
-PROVIDE(GPIO1 = DefaultMachineExternal);
-PROVIDE(GPIO2 = DefaultMachineExternal);
-PROVIDE(GPIO3 = DefaultMachineExternal);
-PROVIDE(GPIO4 = DefaultMachineExternal);
-PROVIDE(GPIO5 = DefaultMachineExternal);
-PROVIDE(GPIO6 = DefaultMachineExternal);
-PROVIDE(GPIO7 = DefaultMachineExternal);
-PROVIDE(GPIO8 = DefaultMachineExternal);
-PROVIDE(GPIO9 = DefaultMachineExternal);
-PROVIDE(GPIO10 = DefaultMachineExternal);
-PROVIDE(GPIO11 = DefaultMachineExternal);
-PROVIDE(GPIO12 = DefaultMachineExternal);
-PROVIDE(GPIO13 = DefaultMachineExternal);
-PROVIDE(GPIO14 = DefaultMachineExternal);
-PROVIDE(GPIO15 = DefaultMachineExternal);
-PROVIDE(GPIO16 = DefaultMachineExternal);
-PROVIDE(GPIO17 = DefaultMachineExternal);
-PROVIDE(GPIO18 = DefaultMachineExternal);
-PROVIDE(GPIO19 = DefaultMachineExternal);
-PROVIDE(GPIO20 = DefaultMachineExternal);
-PROVIDE(GPIO21 = DefaultMachineExternal);
-PROVIDE(GPIO22 = DefaultMachineExternal);
-PROVIDE(GPIO23 = DefaultMachineExternal);
-PROVIDE(GPIO24 = DefaultMachineExternal);
-PROVIDE(GPIO25 = DefaultMachineExternal);
-PROVIDE(GPIO26 = DefaultMachineExternal);
-PROVIDE(GPIO27 = DefaultMachineExternal);
-PROVIDE(GPIO28 = DefaultMachineExternal);
-PROVIDE(GPIO29 = DefaultMachineExternal);
-PROVIDE(GPIO30 = DefaultMachineExternal);
-PROVIDE(GPIO31 = DefaultMachineExternal);
-PROVIDE(PWM0CMP0 = DefaultMachineExternal);
-PROVIDE(PWM0CMP1 = DefaultMachineExternal);
-PROVIDE(PWM0CMP2 = DefaultMachineExternal);
-PROVIDE(PWM0CMP3 = DefaultMachineExternal);
-PROVIDE(PWM1CMP0 = DefaultMachineExternal);
-PROVIDE(PWM1CMP1 = DefaultMachineExternal);
-PROVIDE(PWM1CMP2 = DefaultMachineExternal);
-PROVIDE(PWM1CMP3 = DefaultMachineExternal);
-PROVIDE(PWM2CMP0 = DefaultMachineExternal);
-PROVIDE(PWM2CMP1 = DefaultMachineExternal);
-PROVIDE(PWM2CMP2 = DefaultMachineExternal);
-PROVIDE(PWM2CMP3 = DefaultMachineExternal);
-PROVIDE(I2C0 = DefaultMachineExternal);
+/* This file should be included by hifive1 if virq feature is enabled */
+
+/* Weak symbol for other machine external interrupts handler */
+PROVIDE(OtherMachineExternal = DefaultMachineExternal);
+
+/* Weak symbols for each external interrupt handler */
+PROVIDE(WATCHDOG = OtherMachineExternal);
+PROVIDE(RTC = OtherMachineExternal);
+PROVIDE(UART0 = OtherMachineExternal);
+PROVIDE(UART1 = OtherMachineExternal);
+PROVIDE(QSPI0 = OtherMachineExternal);
+PROVIDE(QSPI1 = OtherMachineExternal);
+PROVIDE(QSPI2 = OtherMachineExternal);
+PROVIDE(GPIO0 = OtherMachineExternal);
+PROVIDE(GPIO1 = OtherMachineExternal);
+PROVIDE(GPIO2 = OtherMachineExternal);
+PROVIDE(GPIO3 = OtherMachineExternal);
+PROVIDE(GPIO4 = OtherMachineExternal);
+PROVIDE(GPIO5 = OtherMachineExternal);
+PROVIDE(GPIO6 = OtherMachineExternal);
+PROVIDE(GPIO7 = OtherMachineExternal);
+PROVIDE(GPIO8 = OtherMachineExternal);
+PROVIDE(GPIO9 = OtherMachineExternal);
+PROVIDE(GPIO10 = OtherMachineExternal);
+PROVIDE(GPIO11 = OtherMachineExternal);
+PROVIDE(GPIO12 = OtherMachineExternal);
+PROVIDE(GPIO13 = OtherMachineExternal);
+PROVIDE(GPIO14 = OtherMachineExternal);
+PROVIDE(GPIO15 = OtherMachineExternal);
+PROVIDE(GPIO16 = OtherMachineExternal);
+PROVIDE(GPIO17 = OtherMachineExternal);
+PROVIDE(GPIO18 = OtherMachineExternal);
+PROVIDE(GPIO19 = OtherMachineExternal);
+PROVIDE(GPIO20 = OtherMachineExternal);
+PROVIDE(GPIO21 = OtherMachineExternal);
+PROVIDE(GPIO22 = OtherMachineExternal);
+PROVIDE(GPIO23 = OtherMachineExternal);
+PROVIDE(GPIO24 = OtherMachineExternal);
+PROVIDE(GPIO25 = OtherMachineExternal);
+PROVIDE(GPIO26 = OtherMachineExternal);
+PROVIDE(GPIO27 = OtherMachineExternal);
+PROVIDE(GPIO28 = OtherMachineExternal);
+PROVIDE(GPIO29 = OtherMachineExternal);
+PROVIDE(GPIO30 = OtherMachineExternal);
+PROVIDE(GPIO31 = OtherMachineExternal);
+PROVIDE(PWM0CMP0 = OtherMachineExternal);
+PROVIDE(PWM0CMP1 = OtherMachineExternal);
+PROVIDE(PWM0CMP2 = OtherMachineExternal);
+PROVIDE(PWM0CMP3 = OtherMachineExternal);
+PROVIDE(PWM1CMP0 = OtherMachineExternal);
+PROVIDE(PWM1CMP1 = OtherMachineExternal);
+PROVIDE(PWM1CMP2 = OtherMachineExternal);
+PROVIDE(PWM1CMP3 = OtherMachineExternal);
+PROVIDE(PWM2CMP0 = OtherMachineExternal);
+PROVIDE(PWM2CMP1 = OtherMachineExternal);
+PROVIDE(PWM2CMP2 = OtherMachineExternal);
+PROVIDE(PWM2CMP3 = OtherMachineExternal);
+/* Weak symbol for I2C0 (g002 only) */
+PROVIDE(I2C0 = OtherMachineExternal);
