+++ src/lib.rs
@@ -12,8 +12,6 @@ pub mod clock;
 pub mod core;
 pub mod delay;
 pub mod device;
-#[cfg(feature = "virq")]
-pub mod exthandler;
 pub mod gpio;
 pub mod pmu;
 pub mod prelude;
@@ -27,5 +25,7 @@ pub mod wdog;
 
 #[cfg(feature = "g002")]
 pub mod i2c;
+#[cfg(feature = "virq")]
+pub mod interrupt;
 
 pub use device::DeviceResources;
